Update getData.py
· Dim = 'All' -> Dim = 0
· Better check of existing/overwriting of files
· No redundancy of print messages

+++ src/nOEN/getData.py
@@ -45,18 +45,22 @@ def createDict(mainKeyName, iDict, info):
         path = '../../Results/'
         fullPathSave = path + info + '.npy'
         if os.path.isfile(fullPathSave) == True:
-            val = input(' > Do you want to overwrite `' + info + '.npy`? [Y/N]: ')
-            if val == 'Y' or val == 'y':
-                np.save(fullPathSave, iDict) 
-                print(' > Data structure saved - `' + info +'.npy`.')
+            cDict = loadResults(info)
+            if cDict['data']['results'] == True:
+                val = input(' > Do you want to overwrite `' + info + '.npy`? [Y/N]: ')
+                if val == 'Y' or val == 'y':
+                    np.save(fullPathSave, iDict)
+                    print(' > Data structure saved - `' + info +'.npy`.')
+                else:
+                    print(' > Data structure not saved. If you don\'t want to overwrite results files, use another name for data Excel.')
             else:
-                print(' > Data structure not saved. If you don\'t want to overwrite results files, use another name for data Excel.')
+                np.save(fullPathSave, iDict)
+                print(' > Data structure saved - `' + info +'.npy`.') 
         else:
             np.save(fullPathSave, iDict)
             print(' > Data structure saved - `' + info +'.npy`.')
     else:
         print(' > No expected Structure: `' + mainKeyName + '`')
-    print('')
 
 
 def loadData(fileName):
@@ -65,65 +69,76 @@ def loadData(fileName):
 
     """  
     print('\n>> Loading data...')
-    # Path name of Data
-    path = '../../Data/'
-    fullPath = path + fileName + '.xlsx'
-    # Initialization 
-    leDict = {}             # Full dictionary (`data` + `comb` + `coeff`)
-    combDict = {}           # Dictionary of combinations (`comb`)
-    coeffDict = {}          # Dictionary of Tau-N coefficients (`coeff`)
-    # Structure dataset (data)
-    di = pd.read_excel(fullPath, sheet_name = 't_0')
-    df = pd.read_excel(fullPath, sheet_name = 't_max')
-    varNames = np.array(df.keys())
-    di = np.array(di)
-    df = np.array(df)
-    numVar = len(varNames)
-    leDict = createDict('data', leDict, {'inocula': di, 'final': df, 'varNames': varNames, 'numVar': numVar})
-    # Structures combinations & Tau-N coefficients (`comb`, `coeff`)
-    v = np.arange(1, numVar+1)
-    nCoef = []
-    for i in range(2, numVar+1):
-        nameK2 = 'D' + str(i)
-        relP = 2/(2**i)     # Reliable point (relP = 2/2^D)
-        ## Combinations (`comb`)
-        c = np.array(list(_comb(v, i)))
-        nc = c.shape[0]
-        nCoef.append(nc)
-        combDict[nameK2] = c
-        ## Coefficients (`coeff`)
-        coeffDict[nameK2] = {}
-        for i_c in c:
-            nameK3_comb = "_".join(varNames[i_c - 1])
-            coeffDict[nameK2][nameK3_comb] = {'iD': i_c, 'D': i, 'reliablePoint': relP, 'numObs': [], 'coeffInfo': {'signs1': [], 'signs2': [], 'deltas': [], 'd_pval': [], 'RKtau': [], 'RKt_pval': []}}
-    combDict['numcoeff'] = nCoef
-    leDict = createDict('comb', leDict, combDict)
-    leDict = createDict('coeff', leDict, coeffDict)
-    # Save structure
-    createDict('saveDict', leDict, fileName)
-    print('>> Loading done.')
+    path = '../../Results/'
+    fullPathFile = path + fileName + '.npy'
+    if os.path.isfile(fullPathFile) == True:
+        print(' > File `' + fileName + '.npy` with results.')
+        rDict = loadResults(fileName)
+    else:
+        print(' > New file .npy.')
+        # Path name of Data
+        path = '../../Data/'
+        fullPath = path + fileName + '.xlsx'
+        # Initialization 
+        leDict = {}             # Full dictionary (`data` + `comb` + `coeff`)
+        combDict = {}           # Dictionary of combinations (`comb`)
+        coeffDict = {}          # Dictionary of Tau-N coefficients (`coeff`)
+        # Structure dataset (data)
+        di = pd.read_excel(fullPath, sheet_name = 't_0')
+        df = pd.read_excel(fullPath, sheet_name = 't_max')
+        varNames = np.array(df.keys())
+        di = np.array(di)
+        df = np.array(df)
+        numVar = len(varNames)
+        leDict = createDict('data', leDict, {'inocula': di, 'final': df, 'varNames': varNames, 'numVar': numVar, 'results': False})
+        # Structures combinations & Tau-N coefficients (`comb`, `coeff`)
+        v = np.arange(1, numVar+1)
+        nCoef = []
+        for i in range(2, numVar+1):
+            nameK2 = 'D' + str(i)
+            relP = 2/(2**i)     # Reliable point (relP = 2/2^D)
+            ## Combinations (`comb`)
+            c = np.array(list(_comb(v, i)))
+            nc = c.shape[0]
+            nCoef.append(nc)
+            combDict[nameK2] = c
+            ## Coefficients (`coeff`)
+            coeffDict[nameK2] = {}
+            for i_c in c:
+                nameK3_comb = "_".join(varNames[i_c - 1])
+                coeffDict[nameK2][nameK3_comb] = {'iD': i_c, 'D': i, 'reliablePoint': relP, 'numObs': [], 'coeffInfo': {'signs1': [], 'signs2': [], 'deltas': [], 'd_pval': [], 'RKtau': [], 'RKt_pval': []}}
+        combDict['numcoeff'] = nCoef
+        leDict = createDict('comb', leDict, combDict)
+        leDict = createDict('coeff', leDict, coeffDict)
+        # Save structure
+        createDict('saveDict', leDict, fileName)
+        rDict = loadResults(fileName)
+    
+    return rDict
 
 
 def loadResults(fileName):
-    print('\n>> Loading results ' + '`' + fileName + '.npy`...')
     path = '../../Results/'
     fullPathSave = path + fileName + '.npy'
     readResults = np.load(fullPathSave, allow_pickle = 'TRUE').item()
-    print('>> Loading done.')
 
     return readResults
 
 
-def writeResults(fileName):
+def writeResults(fileName, Dim = 0):
     """
     - :input:`fileName` (str). Name of file with data and info.
+    - :input:`Dim` (list). List with dimensions we want to write (default: 0 -> 'All').
 
     """
     path = '../../Results/'
     fullPathFile = path + fileName + '.npy'
     fullPathSave = path + fileName + '_results.xlsx'
     if os.path.isfile(fullPathFile) == True:
-        val = input(' > Do you want to overwrite `' + fileName + '_results.xlsx`? [Y/N]: ')
+        if os.path.isfile(fullPathSave) == True:
+            val = input(' > Do you want to overwrite `' + fileName + '_results.xlsx`? [Y/N]: ')
+        else:
+            val = 'Y'
         if val == 'Y' or val == 'y':
             rDict = loadResults(fileName)
             D = rDict['data']['numVar']
@@ -136,7 +151,11 @@ def writeResults(fileName):
             df = pd.DataFrame(data, columns = varNames.tolist(), index = np.arange(1, nd+1))
             with pd.ExcelWriter(fullPathSave) as writer:
                 df.to_excel(writer, sheet_name='Data')
-                for d in range(2, D+1):
+                if Dim == 0:
+                    vD = np.arange(2, D+1)
+                else:
+                    vD = Dim
+                for d in vD:
                     sName = 'D' + str(d)
                     infoR = pd.DataFrame(np.array(['· Dimension ' + str(d), '· Reliable point: ' + str(2/(2**d)), '· Number of observations: ' + str(nd), '· Number of var combinations: ' + str(numComb[d-2])]))
                     infoR.to_excel(writer, sheet_name = sName, index = False, header = False)
@@ -167,5 +186,5 @@ def writeResults(fileName):
 # results = loadResults('data-py')
 # print(results)
 # Write results to Excel
-writeResults('data-py')
+# writeResults('data-py')
 #----------
